refactor: rename gh-actions.nvim to pipeline.nvim
