--- TODO.md
@@ -1,9 +1,0 @@
-# TODO
-
-1. ~Map github responses to internal data structures~
-2. ~Update render to use the new data structures~
-3. Move highlight and icon mapping into provider
-4. ~Move github specific code from `init.lua` into provider~
-5. Implement Gitlab Provider
-6. Rename anything gh-actions related to pipeline.nvim
-7. Rename types from GhActions to pipeline.\*
