+++ lua/pipeline/store.lua
@@ -1,10 +1,10 @@
-local utils = require('gh-actions.utils')
+local utils = require('pipeline.utils')
 
----@class GhActionsStateWorkflowConfig
+---@class pipeline.StatePipelineConfig
 ---@field last_read integer
 ---@field config table
 
----@class GhActionsState
+---@class pipeline.State
 ---@field title string
 ---@field repo string
 ---@field server string
@@ -12,7 +12,7 @@ local utils = require('gh-actions.utils')
 ---@field runs table<integer | string, pipeline.Run[]> Runs indexed by pipeline id
 ---@field jobs table<integer | string, pipeline.Job[]> Jobs indexed by run id
 ---@field steps table<integer | string, pipeline.Step[]> Steps indexed by job id
----@field workflow_configs table<integer, GhActionsStateWorkflowConfig>
+---@field workflow_configs table<integer, pipeline.StatePipelineConfig>
 local initialState = {
   title = 'pipeline.nvim',
   repo = '',
@@ -25,15 +25,15 @@ local initialState = {
   workflow_configs = {},
 }
 
----@class GhActionsStore
----@field package _state GhActionsState
----@field package _update fun(state: GhActionsState)[]
+---@class pipeline.Store
+---@field package _state pipeline.State
+---@field package _update fun(state: pipeline.State)[]
 local M = {
   _state = initialState,
   _update = {},
 }
 
----@param state GhActionsState
+---@param state pipeline.State
 local function emit_update(state)
   for _, update in ipairs(M._update) do
     update(state)
@@ -42,26 +42,26 @@ end
 
 emit_update = utils.debounced(vim.schedule_wrap(emit_update))
 
----@param fn fun(render_state: GhActionsState): GhActionsState|nil
+---@param fn fun(render_state: pipeline.State): pipeline.State|nil
 function M.update_state(fn)
   M._state = fn(M._state) or M._state
 
   emit_update(M._state)
 end
 
----@param fn function<GhActionsState>
+---@param fn function<pipeline.State>
 function M.on_update(fn)
   table.insert(M._update, fn)
 end
 
----@param fn function<GhActionsState>
+---@param fn function<pipeline.State>
 function M.off_update(fn)
   M._update = vim.tbl_filter(function(f)
     return f ~= fn
   end, M._update)
 end
 
----@return GhActionsState
+---@return pipeline.State
 function M.get_state()
   return M._state
 end
