+++ tests/github_spec.lua
@@ -1,5 +1,5 @@
 describe('get_github_token', function()
-  local gh = require('gh-actions.providers.github.utils')
+  local gh = require('pipeline.providers.github.utils')
 
   before_each(function()
     vim.env.GITHUB_TOKEN = nil
